requirements.txt
# FastMCP and MCP dependencies
mcp>=1.0.0
fastmcp>=0.1.0

# HTTP client for API requests
httpx>=0.27.0

# FastAPI for HTTP endpoints
fastapi>=0.110.0

# ASGI server for HTTP transport
uvicorn[standard]>=0.30.0

# Additional utilities
python-dotenv>=1.0.0
